Organize react components
