--- src/App.tsx
@@ -1,11 +1,0 @@
-import { GameDisplay } from './components/GameDisplay'
-
-function App () {
-  return (
-    <>
-      <GameDisplay />
-    </>
-  )
-}
-
-export default App
