+++ src/components/hud/ContractPane.tsx
